requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
datasets==2.18.0
dill==0.3.8
distro==1.9.0
easyocr==1.7.1
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.17.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.4.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyclipper==1.3.0.post5
pydantic==2.7.0
pydantic_core==2.18.1
pytest==8.1.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.13.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tifffile==2024.2.12
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uv==0.1.31
uvicorn==0.29.0
Werkzeug==3.0.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
scikit-learn==1.4.2
jdeskew
openai
groq
pytesseract